+++ src/Utility/Utility.php
@@ -15,8 +15,10 @@ declare(strict_types=1);
 
 namespace DatabaseBackup\Utility;
 
+use Cake\Core\App;
 use Cake\Datasource\ConnectionInterface;
 use Cake\Datasource\ConnectionManager;
+use DatabaseBackup\Executor\Executor;
 use InvalidArgumentException;
 use Symfony\Component\Filesystem\Path;
 use function Cake\I18n\__d;
@@ -29,14 +31,31 @@ use function Cake\I18n\__d;
 abstract class Utility
 {
     public ConnectionInterface $Connection {
-        set (ConnectionInterface|string|null $Connection) {
+        set (ConnectionInterface|string $Connection) {
             if (!$Connection instanceof ConnectionInterface) {
-                $Connection = ConnectionManager::get($Connection ?: 'default');
+                $Connection = ConnectionManager::get($Connection);
             }
 
             $this->Connection = $Connection;
         }
-        get => $this->Connection;
+        get => $this->Connection ?? ConnectionManager::get('default');
+    }
+
+    public Executor $Executor {
+        get {
+            /**
+             * For example, for `Cake\Database\Driver\Mysql` the name will be `MySql`.
+             */
+            $name = substr(strrchr($this->Connection->config()['driver'], '\\') ?: '', 1);
+
+            /** @var class-string<\DatabaseBackup\Executor\Executor> $className */
+            $className = App::classname('DatabaseBackup.' . $name . 'Executor', 'Executor');
+            if (!$className) {
+                throw new InvalidArgumentException(__d('database_backup', 'The Executor class for the `{0}` driver does not exist', $name));
+            }
+
+            return new $className();
+        }
     }
 
     public int $timeOut = 60 {