+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # 2.x branch
 ## 2.14 branch
 ### 2.14.3
+* the `Plugin` class no longer implements the `console()` method (which described the available commands), leaving the
+    application to take care of auto-discovering them;
 * the `AbstractBackupUtility::__call()` method now also supports uppercase named properties (bug fixed). This provides
   the `getConnection()` magic method;
 * the signature of `AbstractBackupUtility::__call()` method has changed (from 
   `__call(string $name, array $arguments = []): mixed` to `__call(string $method, array $args): mixed`);
+* 
 * increased use of fake backup files for testing, where there is no need to use real backups for testing.
 
 ### 2.14.2