+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 
 #### Minor changes:
 * the abstract class `AbstractBackupUtility` has become simply `Utility`;
-* the abstract class `Command` has been moved from `DatabaseBackup\Console` to `DatabaseBackup\Command`.
+* the abstract class `Command` has been moved from `DatabaseBackup\Console` to `DatabaseBackup\Command`;
+* the `@since` tag has been kept for the usual methods that remained unchanged (at least in logic).
 
 # 2.x branch
 ## 2.14 branch